requirements.txt
langchain>=0.2.0
langchain-ibm>=0.3.1
langchain-openai>=0.1.0
langgraph-supervisor>=0.0.7
langgraph>=0.2.33
python-dotenv>=1.0.1
